requirements.txt
# Dev environment
pip
autopep8

# App
google-generativeai # Client library for the Gemini API
Pillow # PIL image-loading library
flask
bidict==0.23.1
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-SocketIO==5.5.1
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.4
pandas==2.2.3
python-dateutil==2.9.0.post0
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2025.2
simple-websocket==1.1.0
six==1.17.0
tzdata==2025.2
Werkzeug==3.1.3
wsproto==1.2.0
pandas>=1.3.0
google-generativeai>=0.3.2